+++ doc/changelog.rst
@@ -9,6 +9,8 @@ Changelog
 * :feature:`162` Test loading and other setup operations now happen before ``session_start``, causing faster failing on simple errors
 * :feature:`-` Log symlinks can now be relative paths (considrered relative to the logging root directory)
 * :feature:`160` Add option to serialize warnings to dicts
+* :release:`0.7.2 <21-08-2014>`
+* :feature:`171` Add error times to console reports
 * :release:`0.7.1 <14-07-2014>`
 * :bug:`-` Fixed error summary reporting
 * :release:`0.7.0 <07-07-2014>`